Move getters and setters after constructor; move "mostPopularName" method in App.java
